Indicate the revision in a content-revision-id header
This makes it easier for clients to extract the MediaWiki revision. The ETag
is not suitable for this, as we'd like clients to retrieve the latest version
after template & image updates. For this reason, RESTBase will return a
timeuuid for the ETag.

We don't use a prefixed header in accordance with http://tools.ietf.org/html/rfc6648.

Change-Id: I75895795561e318ee809760c3d68b9a8101e17a4