Bug 69240: Make lang-links sol-transparent
This solves the empty space problem in [[frwikivoyage:Amérique du Nord]],
for example.

parserTests.txt changed to reflect that language links are not
p-wrapped now. Three tests added to the blacklist because when there
are more than one SOL-transparent links only the last one is taken
out of the <p>. This is being tracked in bug #71361.

Change-Id: I52735d92f0284c3a62b9bbbd25a516d95bf4e998