Use `core-upgrade` instead of require('es6-shim').
The `lib/core-upgrade.js` was introduced as a unified place to
document the exact variant of JavaScript we are targetting -- eg, ES6
plus some helpers like `Array#last` and `prfun` which aren't actually
part of ES6 but we like to use.

Be disciplined to ensure we don't inadvertently have strange bugs:
every entry point (ie, starts with `#!`, is invoked with
`child_process.fork`, or is a module exported as part of the public
API) should immediately `require('core-upgrade.js')` at the top of the
file.  Files which are not entry points should never need to do this.
No file should ever need to `require('es6-shim')` directly.

Change-Id: I71e84e0c7fc8448c19035ef798e2c1c95fa241b8