Strip <meta>/<link> tags while normalizing.
Previously we were stripping these when comparing html/parsoid content
in a parsoid-only test, but not when running a test which is not
parsoid-only.  That doesn't make sense; it should have been the opposite
if anything.  This was causing a bunch of tests written for the php
parser to spuriously fail on Parsoid, because Parsoid was inserting
<meta> tags to record <nowiki>s, <noinclude>s, etc.

This patch makes <meta>/<link> handling consistent whether the test
is parsoid-only or not.

Change-Id: I6034adcb7d7695842c4ba96c1b5d123343d4c902