Chunk-based serializer.
This allow more accurate wikitext escaping by buffering of a line of
wikitext as `ConstrainedText` chunks, which preserve meta-information
about boundary conditions at the edges of the chunks.

For example, the chunk corresponding to a magic link can "remember"
that it needs to have word boundaries on either side.  If these aren't
present (after the chunks on either side have been serialized) then
we can add <nowiki> escapes at the proper places.

Change-Id: I5fad2a24d2aa8e2b11044454a66c7bf5404cbc09