The `tokens` argument to `Util.processTokensToDOM` must be an array.
There are two callers for `processTokensToDOM`:

* `Util.expandValuesToDOM` calls it, after first explicitly testing that
the value it will pass as `tokens` is an array.

* `WikiLinkHandler.prototype.renderCategory` calls it in order to
"deal with sort keys that come from generated content (transclusions,
etc.)".  It passes `content`, which is the result of a call to
`this.addLinkAttributesAndGetContent`.  The control flow of this function
is a little convoluted, but it can be determined by inspection that in
all cases it returns an array.

So we can just get rid of the weird special case handling in
`processTokensToDOM` for non-array tokens.

Change-Id: I18c3daa2cbea778e1d121b95850f2f5accfe0ed5