T51075: Handle template-generated DISPLAYTITLE and DEFAULTSORT.
Use the `properties` key in the action API expandtemplates request so that
we can properly handle DISPLAYTITLE and DEFAULTSORT inside template-generated
content.

Added test cases to parserTests, but they are a bit of cheat: I had to
implement DISPLAYTITLE and DEFAULTSORT in ext.core.ParserFunctions to make
them work, and the tests don't actually exercise the same code path used
in production.

Added a slightly better mocha test, which expands `{{Lowercase title}}` on
enwiki.  This is fragile if enwiki ever changes the {{Lowercase title}}
template.

Change-Id: I459d8b32b6a23f25192042b38a74c8860140e7ec