T71950: Output <nowiki> closer to quotes
Previously, if text contained non-quote escapable chars, it was
fully wrapped in <nowiki>s if necessary, making the tags appear
away from the actual reason for it. With this patch, the code
tries to put them closer to the causing quotes.

Also added some changes for the heuristics to calculate when
two <nowikis> can be merged:
* Ignore display hacks.
* Increase the amount of text allowed inside the same nowiki to 40.

Blacklist changed to reflect new results' outputs. A new test for
nowiki positioning after a ' :' (space followed by colon) was
blacklisted due to parserTests.js detecting some difference when
there shouldn't be.

Change-Id: I576aaf4bcad8931a6cc15d89fb4447125d5f539f