Fix regressions RTing some forms of broken list wikitext
This snippet won't roundtrip
*<small>a
*<small>b

This one won't either:
<small>
*a
*b</small>

But, this one will:
<small>
*a
*b
</small>

The last one will because formatting tags are outside the list and
we have special support for it in our wt2html pipeline (see 600c60b8).
However, the first two are more complex because the formatting markup
is now embedded without the list elts and we let the HTML tree builder
to fix it up.

* This patch updates the li-handler cleanup from 14c47443 to handle that.
* Added / updated parser tests and blacklist for tests.
* Need to revisit one of the tests to see if we want to change
  Parsoid's wt2html output on it.
* Fixes RT-testing regressions.

Change-Id: I71de8faa2e05ec01e13986c618ada362632fef53