T73708: Serialize category redirects with a ':'
Redirects to a category should be serialized with a ':' prefix so
the page isn't also added to the category. Unless there's also a
real link to the category just after, in which case we can merge
the two into "#REDIRECT [[Category:…]]".

Also fixed a bug where all category redirects where parsed as both
redirects and category links, even if they started with ':'.

Change-Id: I504ff21efa494f1d5964208af9fa0f4cdcf5e198