[parserTests] Ensure manual selser changes have a clean DOM
The test item for manual selser changes needs to be cloned from
the original item when --selser=noauto is activated from the
command line so that its DOM is not clobbered by a previous test.

Also, avoid directly changing the changes list in
applyManualChanges as that can affect following tests. To test,
pass both --wt2wt and --selser on the command line.

Change-Id: Ib8de2c8bb3b427fb31d3c1c95ddd67af3042bf83