Don't strip id attributes from DOM nodes -- required for <ref> tags
While we've stopped corrupting pages with today's deploy, we are
now returning http 500s in scenarios where because of edits, some
<ref>s get marked modified by the DOMDiff algorithm. This bug
didn't affect new <ref> tags. Because of the unsophisticated nature
of our DOMDiff algorithm, sometimes, wrappers of unmodified
content do get marked modified and if an unedited <ref> happened
to be one of those, the Parsoid serializer tripped on this bug.

Bug: T93228
Change-Id: Ia9477057e13c5831f08464a44ac816b1c080b9da