T95039: encode/decode arbitrary data in comments.
Use HTML5 entity encoding to ensure that arbitrary data can be represented
in comments without triggering premature end of comment.

Do a little bit of entity encoding on the wikitext side as well to allow the
user to author the string "-->" inside a comment.  They probably won't want
to (!) but we want to avoid exposing arbitrary content restrictions to
DOM consumers (like Visual Editor).

If you were wondering, "-->" is represented as "--&gt;" in wikitext.

We also normalize HTML entities during DOM diff, which ensures that
we don't dirty diff even if VE does the entity encoding inside comments
differently than we do.

Change-Id: Id1cb60d201abfc14633a353830ea9fbc89f633a8