Exponentially increase the request timeout
The Parsoid API request client so far operated with a fixed timeout, which in
the case of parse requests was only 16s. For large gallery tags parse times
could exceed 16s on each attempt (probably ~40s), which caused Parsoid to
retry up to eight times before giving up.

This patch increases the base timeout for parse requests to 30s, and
multiplies the timeout by a factor between 3 and 4 on each retry. The exact
factor is randomized, so that retries don't happen in lock-step. With this in
place, even huge gallery tags should be successfully parsed on the first retry.

Bug: T97204

Change-Id: I8798d51c530dc438be890e022cbe82d2dcc6345c