T73161: Support subst: of transclusion blocks
Add new API parameter 'subst'. If set, top-level templates passed
in wikitext will be returned with just their transclusion parsed
into HTML, instead of the full (all templates) transclusion.

As the result is stripped of data-parsoid (so that the content is
considered new), the option is only supported for the HTML format,
not for pagebundles.

Also added a very simple implementation of onlypst to the mock
MediaWiki API, which only subst's the {{echo}} template.

Change-Id: Id1c2bf8e9e4af0e324968fa3096b9217b1e0cdcf