Allow `rtTestMode` option to be inherited via the prototype chain.
This is just a code style thing -- the `hasOwnProperty` call here is
weird.  Do the standard thing: use `undefined` to mean not set, and
`true` and `false` for the set values.  And if someone decides someday
to clone and tweak options via `Object.create(options)`, they won't be
unpleasantly surprised when `rtTestMode` works differently than every
other option property.

Change-Id: I9573865655824e6223c8e9cfdac57731cf9e4bdc