Replace most instances of "production" with "rule"
In Bryan Ford's 2002 thesis, the term "rule" is used, and in his 2004
paper, the term "definition" is used. PEG.js uses "rule" so I will
follow its example here. Plus it is a nicer word.

In neither of Bryan Ford's seminal papers are rules ever called
productions. The reason is explained in the 2002 thesis in section
2.3. A CFG tells you how to write a string, whereas a PEG tells
you how to read a string. Thus in the CFG terminology, we have a
grammar "producing" a string, and such productions are written with a
rightwards arrow, whereas in a PEG we have a "rule" for recognising a
string, written with a leftwards arrow.

Change-Id: I2ecbc7c92baaf95955dec0ac0390df9adac6b06f