In the tokenizer use the "$" operator
The "$" operator returns the matched input text, which is useful in many
cases. It avoids the need to do a join(''), and often can be used to
ignore the undefined match result returned by assertions, e.g.

!x c:. {return c;}

becomes

$(!x .)

which is more efficient since the former creates a small JavaScript
function in the generated output which does nothing except pass through
its first argument.

Change-Id: I3b6b1a4214c143d84a856195fa7038b890f815f7