parserTests.txt: Use a light touch when normalizing html/parsoid.
Heavy normalization is warranted when we're trying to match html/php
output with Parsoid output in parserTests.  But when the parserTest
is an html/parsoid section, we want to leave it pretty much alone
so that we can accurately check empty <span>s, thumbnail paths, URI
encoding, etc.

In fact, tweaking the normalization turned up a legit bug in how
the `src` attribute of `<img>` tags was encoded.  This patch fixes
this minor issue as well.

Note that I still strip <meta> tags in html/parsoid sections.  This
is because we frequently test WTS <nowiki> insertion by providing
an html/parsoid section with the <meta> tags stripped out, allowing
the html2wt test to verify that the <nowiki> is correctly added,
while still allowing the html2html and wt2html versions of the test
to pass.  If <meta>s were not stripped, these tests would all have
to be modified and split up.  Not worth it at this time.

Change-Id: Ib6f61ba81351e39246f5cdbf1500b82407564328