Make tokenizer errors be more vague
It's very expensive to collect error position and expectation
information: it requires a call to peg$computeLocation() on every
subexpression failure, just in case the whole start rule later fails.
Wrapping the start rule in a "named" node causes this subexpression
failure collection mechanism to be suppressed, improving speed by 30% on
top of the previous no-cache patch, for the same test case.

Change-Id: I6fd65fed2476177046e1283085be88ad07a1cf20