T104918: Fix redirects to non-local targets.
Avoid duplicating the link-handling code: defer processing of redirect links
to the normal link handlers, and convert the resulting href (even if non-local)
into a <link rel="mw:PageProp/redirect" href="...."/> tag.

The results of the redirect-related parser tests have also been compared to
the PHP implementation, and some discrepancies corrected.  In particular,
interwiki links are valid redirects (although the redirect is not
automatically followed), and `#REDIRECT [[<nowiki>[[Bar]]</nowiki>]]` is
*not* a valid redirect.  Some namespaced links have also been prefixed
with `./` to avoid parsing the namespace as a URL protocol.

Language links are treated as valid redirects.  This is slightly
different from how they are handled in PHP (see T104918), but we
believe the PHP behavior to be a bug.

Change-Id: If077218126e3608e2e9efd15fe54871a7ead169b