doc: jscs @return markup
This rule might be too flakey to enable; the code analysis it uses appears
to be insufficent to handle some interesting object return cases in our
code base.  On the other hand, it's easy to turn off on a case-by-case
basis where needed, and it did catch a few bad @return types.

Change-Id: I6c10654d1bea9b86e713a896d954dc95d7769539