T109371: Protect RFC/PMID/ISBN magic links with <nowiki> during WTS
The old magic link logic in wts.escapeWikitext seemed to have
bitrotted, and was always evaluating to true (causing magic links
never to be nowiki-protected).  Fix the logic to be consistent with
the logic for protecting autolinks.

Note that currently this routine is only triggered when state.inLink
is false, since that's part of the "fullCheckNeeded" test in
WEHP.escapeWikiText.  But add state.inLink to the test in
`WEHP.hasWikitextTokens`, just in case some other code path or future
refactoring brings us there.

Note that this patch adds a few new <nowiki>s to the blacklist,
which is a consequence of a bug in Parsoid's handling of whitespace
in magic links, addressed by I57033647c5084cf281d735c6f245a5b98a235b38.

Change-Id: I5dd7000656ba4ca9711779f9b677eb6879735214