Quote the profile in our content type strings
According to RFC 7231 section 3.1.1, parameter values are either tokens
or quoted strings.  The forward-slash character ("/") is not allowed in
a token (defined in RFC 7230, section 3.2.6).  We need to make the
parameter value a quoted string.

This parsing restriction is actually enforced by the npm `content-type`
module (version 1.0.1) which is used by Express 4.x to manipulate
content type headers.  The unmodified content-type causes this
exception:

  TypeError: invalid parameter format
    at Object.parse (/home/cananian/Projects/Wikimedia/Parsoid/node_modules/express/node_modules/content-type/index.js:140:13)
    at setCharset (/home/cananian/Projects/Wikimedia/Parsoid/node_modules/express/lib/utils.js:275:28)
    at ServerResponse.send (/home/cananian/Projects/Wikimedia/Parsoid/node_modules/express/lib/response.js:154:32)

when we try to use `res.send` (for example, in `apiUtils.sendResponse`)
after setting the content type to `WIKITEXT_CONTENT_TYPE` or
`HTML_CONTENT_TYPE`.

Change-Id: I4a22830248b712f340c60a9f417395a6ac90fb76