Restore mangled + clean up all wikitext escaping tests
* nowiki spans should not be present in the html/parsoid
  sections where we are testing our nowiki functionality.
  We want nowikis to show up without it being present
  in the html.

* Reorganized tests so that html/parsoid shows up first,
  followed by wikitext and then by html/php, if any.

* Some tests have html/php (directly or via html/*) sections
  -- the idea was to verify that our nowiki-escaped output
  parses to html that looks reasonable compared to what we
  started out with. Not all tests have this. The utility of
  this is unclear, but I am leaving behind those that have them.

* We are content doing html2wt on this. There are ample tests
  for other test modes elsewhere in the test file. So, added
  parsoid=html2wt to the options section for all these tests.

Tangent:
-------
We need to start thinking about overhauling the parser tests
infrastructure and split this humongous file into multiple
files, and organized by functionality and what it tests.

We can introduce include directives if necessary to make sure
that the tests still run as a monolothic entity, it that is
desirable.

But, this file is getting unwieldy and it gets hard to detect
undesirable / unwanted edits to tests. Will create a phab task
for it and write up some thoughts about it.

Change-Id: I2c2d78ebbbc1d5f1e1b413f4e526f33210b7c511