Fix denial of client-side upscaling in thumb and frameless format
The batch API returns the input width and height even when
client-side scaling is done, which I think is the correct policy.
However, this does mean that Parsoid needs to handle the case where
upscaling is denied itself, instead of relying on the API.

Also, mediatype === 'DRAWING' was used as an approximation to
mustRender(), which is not accurate. For example, PDFs have media
type "OFFICE", see MW's includes/mime.info. Use mediatype != 'BITMAP'
in the b/c case, which is more likely to be correct. Also
allow the API to provide explicit indication of mustRender. A patch
will be forthcoming implementing this in the batch API. This allows
Parsoid to match the behaviour of MW, which uses mustRender to
determine whether upscaling should be allowed.

Bug: T112668
Change-Id: I1500c0f4623928c3be6726cddc33cc31b2b9e00d