Make `parserTests.js --quiet` truly quiet
Add a `suppressErrors` option to the parsoid clause in parserTests to
suppress error messages when they are expected (for example, when
testing error handling in Parsoid).  Also suppress errors for
tests of <gallery>, since we know that doesn't work yet.

Tag tests with `suppressErrors` where needed, and make a few others
PHP-only.  Count and report the number of errors logged during
parserTests execution, and exit with a non-zero code if we log
errors.  This ensures that jenkins will help us keep parser tests
quiet.

Change-Id: I7593d579d9487df69cc9a5f64efe2efa36c7668e