Disable migrateTrailingNLs if table has had content fostered out of it
Added test cases for T115289 which illustrate the cases we've fixed.

But also added a new test case for T73074 which shows that some cases
with templates inside tables still lead to duplicated content.  What
needs to happen here is the template-affected region needs to be
expanded further to include the fostered content.

Bug: T115289
Change-Id: I0047a5e7ccb067b7e40d04d280bafb862abcfba5