In parse.js, send a message to stdin if there is no input for 1s
The convention that empty input on stdin causes a remote request is
confusing. I failed to learn the first time and just did it again, so
I'll take that as evidence that it's a common error. So, send a message
to stderr warning the user that parse.js is waiting for input.

Change-Id: I8248f7b603afd6bcc2b1dce90d80a4b30202ea1e