T118367: Kill dead code + fix bad perf in pathological scenarios.
This code was added in fd0aefc3 (Aug 3, 2012) in the very
first draft of the template encapsulation code when nothing
but the very outlines of this technique had been dreamt up.

This code was seemlingly added to deal with fostered content from
tables (suprise!). But, given how our template wrapping works,
about tags are only assigned to the start/end meta tags and
nothing else in between. So, there was no way that a table
could exist with an about-id found on transclusion meta tags.

You can verify this by examining code in tt/TemplateHandler.js
and/or by running this:
echo "{{echo|a\n\nb\n\ncr\n\nd}}" | node parse --dump dom:pre-encap
and noticing that none of the p-tags there have an about id.

So, this code has always (or at least since almost the very
beginning) been dead, and only served to make parsing slow in
pathological cases as in the bug report.

With this dead code removed, that page parses in 17s.

Change-Id: I056169adab7f958f35c2ecb84135c61a4736e9b1