Bump html content-type version to 1.2.0
There have been some changes to the HTML that Parsoid emits
since the last deploy.

* <meta property="mw:parsoidVersion" content="0"/> is now changed to
  <meta property="mw:html-content-type" content='version-string'/> to be more
  consistent with the version information that is output in the response
  headers.

* For the non-pagebundle API endpoints,
  <meta property="mw:data-parsoid-content-type" content='version-string'/>
  is emitted.

* data-mw.body.extsrc is no longer emitted for the <references> tag.

* User information is no longer emitted in the <head> tag as per T125266.

While Parsoid's serialization is backward-compatible with this change
and all known clients aren't affected by these changes, bumping the
version number is good practice and also lets us isolate HTML changes
in storage and purge this HTML from storage (for whatever reason or if
we need to revert the deploy).

Change-Id: I792867f4846676bd1c19541ec0fb9a070e0af6e8