Config: expect "port" and "interface" as config option names
Parsoid used to have serverPort and serverInterface as the configuration
option names for the port and interface, respectively. However, in WMF
production, services based on service::node get the "port" and
"interface" config options instead. This commit changes that, all the
while keeping the compatibility with previous config versions so as not
to affect existing instalments.

Bug: T90668
Change-Id: Icd20f7cc63e9892fcd5f1af85c947f803b903200