migrateTrailingNLs: Process children last -> first
Since we have code checks to suppress newline migration
within tables with fostered content, we need to process
a table before its fostered content is processed to prevent
the fostering connection being broken by a newline migration
from fostered content.

The blacklist changes are because of a change in newline
migration on the table.

Change-Id: Ia2edf4760c279838a1a84804fd443766333dd55d