linter: Implement sampling options
When enabling this in production, we want to be very careful not to
overload MediaWiki with a large number of lint errors. To do this, there
are two main linting features:

First, enable selective categories to be reported to MediaWiki. The
"linting" configuration setting can be a list of the names of lint
errors to enable, or true to enable everything.

Then is a per page sampling feature. Since we also log deletions, we
need this to be deterministic, so we use page_id % sampling === 0 to
check whether logs should be sent.

Change-Id: I2f939859a6f49b40aff93f3c07b11c64c0b00fc1