Add a mwApiServer configuration variable
In the current setup, we cover the case in which the connection to the
MediaWiki API is performed through a proxy; we don't consider the case
in which we want to connect to a specific IP to serve all the
MediaWiki requests. This is useful when the MediaWiki installation must
be reached on an internal service IP rather than its public
endpoint. Historically this has been achieved by setting the IP/hostname
as the proxy URI, but that only worked correctly as long as we don't
need to support TLS.

Change-Id: I32d041fce6707e8e14546395e26ffa5a690c2862