Parse and serialize language converter markup.
This should parse all valid language converter markup and turn it into
DOM.  WTS functionality is implemented.  It doesn't implement the
post-processing step that actually performs the language conversion.

The markup uses the `data-mw-variant` attribute because these
properties are needed for rendering (esp if the language conversion is
done client-side).  We can't use the existing `data-mw` attributes
because they are suppressed for rendering (and, to the point, inside
templates).

There one significant difference with PHP parser behavior.  As
mentioned in https://phabricator.wikimedia.org/T43716#469890, language
converter markup which overlaps DOM tree boundaries is clipped to DOM
tree boundaries.  For example:

  -{zh-cn:<span style="color:red">;zh-tw:<span style="color:green">}-foo</span>

when piped through

  bin/parse.js --domain zh.wikipedia.org --normalize=parsoid

becomes:

  <p><span typeof="mw:LanguageVariant" data-mw-variant='{"bidir":[{"l":"zh-cn","t":"&lt;span style=\"color:red\" data-parsoid=&#39;{\"stx\":\"html\",\"autoInsertedEnd\":true,\"dsr\":[8,32,24,0]}&#39;>&lt;/span>"},{"l":"zh-tw","t":"&lt;span style=\"color:green\" data-parsoid=&#39;{\"stx\":\"html\",\"autoInsertedEnd\":true,\"dsr\":[39,65,26,0]}&#39;>&lt;/span>"}],"show":true}'></span>foo</p>

Note that the open `<span>` inside the language converter block is
automatically closed, and does not include the `foo`.  This is similar to
the "balance" constraints we would like to impose on template inclusions
(T114445).

In theory this could even occur on wikis where language converter is
disabled but a wikitext construct spanned the `-{ ... }-` block.
In practice, because of the preprocessor brace-matching regularized by
https://gerrit.wikimedia.org/r/333997 and because HTML tree-building
occurs after tokenization, differences should be rare. For example,
if you change the domain to `en.wikipedia.org` in the example above,
you get `<span>` tags spanning `-{...}-` as you'd expect:
  <p>-{zh-cn:<span>;zh-tw:<span>}-foo</span></span></p>

Bug: T43716
Depends-On: I2f0c186c75e392c95e1a3d89266cae2586349150
Depends-On: I65736520cd04bfe8949b29ade07338a6e1b88a4d
Change-Id: I4fcdebc2290ec35ba188f4c2e69d578791fbcd67