Update baseconfig; add frwiki.
Add a README which explains how to regenerate the baseconfig JSON files
using tools/sync-baseconfig.js.

Regenerate all baseconfigs *except for enwiki* from the current wiki
configurations.  Add frwiki, since this is needed in an upcoming
parserTests merge.

The enwiki configuration appears to have changed in a way which causes
numerous parserTests failures (T156295, T156296).  So we haven't updated
it here; we'll tackle those issues in a follow-up patch.

Change-Id: Id36550e864acf0df472b6a6a12f750eb28cbdd1d