Implement {{PAGELANGUAGE}}; fix language code in HTML <body>
Parsoid had been naively using wiki language, not the page-specific
language.  We need to fix this for LanguageConverter support, since
conversion is enabled based on whether the *page language* has variants.

As suggested by https://phabricator.wikimedia.org/T59603#2970423,
you can test this with:

 $ bin/parse.js --page Bot/es --domain meta.wikimedia.org < /dev/null

which should return `lang="es"` as an attribute on the `<body>` tag.

Bug: T59603
Change-Id: I26d7972a54a214a17bb92d735aa41f1b030a38d6