Address edge case in DSR code that affects lint output in some cases
eccca398 added some DSR correction code for b/i tag misnesting
by (correctly) swallowing the DSR width of the stripped meta tag.

However, this also causes the stripped-elt meta tag itself to get
its DSR corrected to zero-width because of the forward pass.

This is not a problem normally since the meta tag will get stripped.

However, this trips up the linter stripped-tag DSR values for snippets
with i/b tags in them. It emits zero-width DSR for them. To prevent
this, we record the original DSR in a dp.tmp and utilize it in the
linter.

Added a new test case that passes with this patch and fails without.

Change-Id: Ie15c5311957995bf4fdbb97771e927b6ac39ba9e