Preprocessor precedence: rightmost-opening/"broken-link"/"broken-template"
Match the behavior of the PHP preprocessor for links or templates missing
their close brackets or misnested, by using a shared stack for brackets in
inlineBreaks.  We only stop for the close-symbol of the rightmost opened
construct; see [[:mw:Preprocessor_ABNF]] for details.  So:
  [[Foo|bar}}]]
and
  [[Foo}}|bar]]
are both valid, even if there's an template opening somewhere to the left.
Note however that the latter example will not become a link in the output
because }} is not a valid title character -- it is unlinked at a later
stage.  Similarly:
  [[Foo|{{echo|]]}}
is a broken link, because the square brackets are not inline breaks inside
the template.

Note that there is a deliberate incompatiblity with PHP.  Because PHP has
a separate preprocessor stage that recognizes *but does not mark*
wikilinks, markup like:
  [[Foo{{echo|]]}}
or
  [[Foo{{echo|]]
is parsed as a broken wikilink in the PHP preprocessor, but it then emits
an intermediate result (`[[Foo]]` or `[[Foo{{echo|]]`) which is then
later parsed as a wikilink by the PHP parser.  This is arguably a bug,
which could be fixed by having the PHP preprocessor either explicitly
mark wikilinks, or else HTML-escape `[[` when it appears in the
`broken-wikilink` production. A rough sketch of a PHP-side fix can be
found in I0dd4512dc9014fc8e329e682be20184867cafe80.

Change-Id: I8a2dd4471383b389dc17fefb7917305aefcb3141