T114072: Always emit a lead <section> tag
This is a follow-up to I0f4c19f7ca4ebc88eb85c53ab145b54fde7ab455 to
tweak the spec such that there is always a "lead section" generated,
even if it would consist only of whitespace and comments.  This avoids
the need for special cases in VE (when it displays comments in a lead
section) and the editing interface (since PHP gladly accepts an edit
to "section 0", even if there is no content in the lead section in
the existing page).

Tweaks the parsertest blacklist, due to the selset newline diffs
described previously in I0f4c19f7ca4ebc88eb85c53ab145b54fde7ab455.

Change-Id: Idf198a871c2d416ac367182cb5ded1b8872a1b59