Fix crasher in getDSR in section wrapping code
When resolving section-template conflicts, we try to expand the range
of template ranges to includes all content within a section. As part
of that, we need to update the DSR and traverse the non-template
content to compute offsets. When non-templated content only includes
text and comment nodes, we hit templated content. At that point, we
should get the baseline DSR info from the template wrapper. Otherwise,
if we query the last node of the templated content region for DSR
info, we get an undefined value.

This should fix crashers found in RT testing.

I left behind section counters added during debugging.

Change-Id: I8eb21996b9e7bcb856322a7bdfa6eec5882303a0