Use a Set for trace/debug/dump flags.
This reads more naturally than repeated indexOf() operations against
an array.  We pass around the Parsoid options through inter-process
communication with workers, and whatever we send through that pipe
needs to be JSON-serializable.  So Util.setDebuggingFlags() still
creates arrays, and we pass arrays in the parsoidOptions configuration
object, but create Sets from them in the ParsoidConfig constructor.

Change-Id: I6531dda9d525702c76e4fe1f2911dea9a665b270