Make the log levels configurable for the initial logger
To avoid having to reconstruct and set a logger at a reduced level, as
in parserTests.

Set the default for bin/parse to "warn".  The necessitates bumping the
level for linting (when not sending to the MW API) so that --linting
reports the errors w/o needing the newly added --verbose, which restores
the "info" level.

Change-Id: I2d04edc15912b0ce73285ab7d9c4fc7cf5f9b457