Only use max depth detection since we aren't doing template expansion
Testing this in parserTests runs afoul since we're still using native
template expansion there and we hit loop detection first.

Locally, one can try a Template:Gallery,

<gallery>
File:Test.png|123 {{Gallery}}
</gallery>

which renders similarly to the php parser, but with 2x instances.

References are a special case though, which we'll see in the follow-up.

Change-Id: I1a026191c4246180c210ad385a624fe9a6b2d2f6