Account for SOL transparent templates in p-wrapping
Fixes the gratuitous template wrapping on,
/en.wikipedia.org/v3/page/html/Wikimedia_Foundation/803552163

This change allows the templated category to be left out of p-wrapping
so that the only thing on the line that is wrapped after the closing div
is the closing span.  The paragraph gets removed as useless and the span
then returns to the same depth as the opening tag, which avoids wrapping
the entire page.

Bug: T182349
Change-Id: I5093a3af83db2352309c2a39e95412c441cf68aa