Use Promise.async/yield in roundtrip-test.js
This is a clean up in advance of cloning this for the language converter
test facility.  Audited code to ensure that the legacy 'callback' parameter
to runTests was never used, and removed the weird corner case where passing
'check' in the runTests options would change the structure of the return
value.  Also removed our usage of the deprecated `--prefix` option in
the `roundtrip` test in package.json, and added `--domain` options where
they were previously implicit, just to make our test cases explicit.

Depends-On: I732ae30a5295f6b1f833553c2293bec8b13ff1c9
Change-Id: I8cad43c9ae0e6046957c57b9233c69be2f7c0473