WTS: Use `yield` instead of directly returning Promises
Direct return of Promises from async functions, although convenient,
can obscure the callstack if an exception is thrown (since the caller
is no longer on the callstack when the returned Promise is resolved),
and may confuse readers which have to keep track of whether the return
value is of Promise type or not in order to figure out control flow
in the case of an exception.

Prefer to explicitly `yield` rather than directly returning a Promise,
at least in the WTS code.

Change-Id: I88d5b628d3e0ac0a208f6d8188ba23b78c8ce03a